cli/requirements.txt
appdirs==1.4.3
astroid==2.3.3
atomicwrites==1.3.0; sys_platform == "win32"
attrs==19.3.0
bandit==1.6.2
black==19.10b0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.0
codecov==2.0.15
colorama==0.4.3; sys_platform == "win32" or platform_system == "Windows"
coverage==5.0.3
dnspython==1.16.0
email-validator==1.0.5
gitdb2==3.0.2
gitpython==3.0.8
google-api-core==1.16.0
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-datastore==1.10.0
google-cloud-firestore==1.6.2
googleapis-common-protos==1.51.0
gql==0.4.0
graphql-core==2.3.1
grpcio==1.27.2
grpcio-gcp==0.2.2
idna==2.8
importlib-metadata==1.5.0; python_version < "3.8"
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mock==4.0.1
mock-firestore==0.6.0
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
password-generator==0.1.0
pathspec==0.7.0
pbr==5.4.4
pluggy==0.13.1
promise==2.3
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pyparsing==2.4.6
pyperclip==1.7.0
pytest==5.3.5
pytest-cov==2.8.1
pytest-mock==2.0.0
python-gnupg==0.4.5
pytz==2019.3
pyyaml==5.3
regex==2020.1.8
requests==2.23.0
requests-mock==1.7.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
rx==1.6.1
semver==2.9.1
six==1.14.0
smmap2==2.0.5
stevedore==1.32.0
tabulate==0.8.6
toml==0.10.0
tqdm==4.42.1
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
zipp==2.2.0; python_version < "3.8"
